Initial commit. It's a dog's breakfast!
Problems at the moment are in filegrabber.rb - I gather the way to download attachments with html files is to use a :chrome driver, but I can't install capybara-webkit because it won't find qtwebkit on my system even when I manually isntall it! Aarrggh. And I don't know how to use Selenium, and the methods I was using with Capybara (session.visit $URL) it says 'method not found'. The examples online are incomplete...